requirements.txt
asgiref==3.3.4
Django==3.2.4
django-cors-headers==3.7.0
django-localflavor==3.1
djangorestframework==3.12.4
localflavor==1.9
psycopg2==2.8.6
python-dotenv==0.17.1
python-stdnum==1.16
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
whitenoise==5.2.0
